Enforced Code Style & Added support for SDL3_Mixer

+++ Graphs/HeuristicFunctions.h
@@ -8,47 +8,46 @@
 /*=============================================================================*/
 
 #include <algorithm>
+#include <cmath>
 
 
 namespace bae::Graphs::HeuristicFunctions
 {
-#pragma warning(push)
-#pragma warning(disable : 4505)
-	//Manhattan distance
-	static float Manhattan(float x, float y)
-	{
-		return static_cast<float>(x + y);
-	}
+    #pragma warning(push)
+    #pragma warning(disable : 4505)
+    // Manhattan distance
+    static float Manhattan(float x, float y)
+    {
+        return static_cast<float>(x + y);
+    }
 
-	//Euclidean distance
-	static float Euclidean(float x, float y)
-	{
-		return static_cast<float>(std::sqrtf(x * x + y * y));
-	}
+    // Euclidean distance
+    static float Euclidean(float x, float y)
+    {
+        return static_cast<float>(std::sqrtf(x * x + y * y));
+    }
 
-	//Sqrt Euclidean distance
-	static float SqEuclidean(float x, float y)
-	{
-		return static_cast<float>(x * x + y * y);
-	}
+    // Sqrt Euclidean distance
+    static float SqEuclidean(float x, float y)
+    {
+        return static_cast<float>(x * x + y * y);
+    }
 
-	//Octile distance
-	static float Octile(float x, float y)
-	{
-		//const float f = 0.414213562373095048801f; // == sqrt(2) - 1;
-		const float f = std::sqrtf(2) - 1; // std::sqrt is constexpr in C++26
-		return static_cast<float>((x < y) ? f * x + y : f * y + x);
-	}
+    // Octile distance
+    static float Octile(float x, float y)
+    {
+        //const float f = 0.414213562373095048801f; // == sqrt(2) - 1;
+        const float f = std::sqrtf(2) - 1; // std::sqrt is constexpr in C++26
+        return (x < y) ? f * x + y : f * y + x;
+    }
 
-	//Chebyshev distance
-	static float Chebyshev(float x, float y)
-	{
-		return std::max(x, y);
-	}
+    //Chebyshev distance
+    static float Chebyshev(float x, float y)
+    {
+        return std::max(x, y);
+    }
 
-#pragma warning(pop)
-
-
+    #pragma warning(pop)
 }
